AI Sales/requirements.txt
openai
pinecone-client
streamlit
streamlit-chat
pandas
numpy
pytube
python-dotenv
flask
flask_mysqldb
passlib
pandasql
openpyxl